pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "mcp>=1.3",
    "qdrant-client>=1.12",
    "sentence-transformers>=3.0",
    "python-frontmatter>=1.1",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "pytest-mock>=3.14"]

[tool.setuptools.package-data]
memory_mcp = ["templates/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
